MCR-1694 fix logging, move util class
