+++ mycore-base/src/main/java/org/mycore/frontend/cli/MCRDerivateCommands.java
@@ -69,7 +69,7 @@ import org.mycore.datamodel.metadata.MCRObjectID;
 import org.mycore.datamodel.niofs.MCRAbstractFileSystem;
 import org.mycore.datamodel.niofs.MCRPath;
 import org.mycore.datamodel.niofs.utils.MCRTreeCopier;
-import org.mycore.frontend.MCRDerivateUtil;
+import org.mycore.datamodel.niofs.utils.MCRDerivateUtil;
 import org.mycore.frontend.cli.annotation.MCRCommand;
 import org.mycore.frontend.cli.annotation.MCRCommandGroup;
 import org.xml.sax.SAXException;
@@ -913,8 +913,7 @@ public class MCRDerivateCommands extends MCRAbstractCommands {
 
     @MCRCommand(syntax = "test rename file {0} with {1} to {2}",
         help = "TODO please add help text")
-    public static void testRenameFile(String filename, String pattern, String newName)
-        throws IOException {
+    public static void testRenameFile(String filename, String pattern, String newName) {
         MCRDerivateUtil.testRenameFile(filename, pattern, newName);
     }
